requirements.txt
asgiref==3.5.0
astroid==2.11.4
autopep8==1.6.0
bidict==0.22.0
colorama==0.4.4
coverage==6.3.2
dill==0.3.4
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
eventlet==0.25.2
Faker==13.6.0
greenlet==1.1.2
importlib-metadata==4.11.3
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.7.0
model-bakery==1.5.0
monotonic==1.6
Pillow==9.1.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
Pygments==2.12.0
PyJWT==2.3.0
pylint==2.13.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
Unidecode==1.3.4
wrapt==1.14.1
zipp==3.8.0